pyproject.toml
[project]
name = "typervox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "faster-whisper>=1.2.0",
    "pyaudio>=0.2.14",
    "sounddevice>=0.5.2",
    "python-xlib>=0.33",
]

[project.scripts]
typervox-cli = "typervox.cli:main"

[tool.uv]
package = true
